% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotEnVariabel.R
\encoding{UTF-8}
\name{barplotEnVariabel}
\alias{barplotEnVariabel}
\title{Barplot af en variabel fra et datasæt}
\usage{
barplotEnVariabel(
  datasaet,
  xVariablen,
  yVariablen,
  overskrift = NA,
  navnPaaXAksen = NA,
  navnPaaYAksen = NA,
  inputTilLegend = NA
)
}
\arguments{
\item{datasaet}{en data.frame, Datasættet med de variable som skal bruges til plottet.}

\item{xVariablen}{en character string, Navnet på x-variablen i datasaet}

\item{yVariablen}{en character string, Navnet på y-variablen i datasaet}

\item{overskrift}{en character vector af en maks. længde på 4 eller NA, Overskriften på plottet. Hvis NA så vil overskriften sættes på ud fra xVariablen og yVariablen}

\item{navnPaaXAksen}{en character string eller NA, Navnet på x-aksen}

\item{navnPaaYAksen}{en character string eller NA, Navnet på y-aksen}

\item{inputTilLegend}{en chracter vector, ekstra informationer/tekst som man ønsker at have på som legend. Hvis NA, vil der ikke komme legend på grafen.}
}
\value{
En barplot med eventuelt legend.
}
\description{
Barplot af en variabel fra et datasæt
}
\details{
yVariablen i datasaet skal være af typen numeric/integer.
}
\examples{
antalPatienterDat <- data.frame("maaned" = month.name[1:12], "antal" = sample(100:150, 12))
barplotEnVariabel(datasaet = antalPatienterDat,
                  xVariablen = "maaned",
                  yVariablen = "antal",
                  overskrift = "Antal patienter med x sygdom i 2018" ,
                  navnPaaXAksen = "Maaned for besoeg paa afdelingen",
                  navnPaaYAksen = "Antal patienter",
                  inputTilLegend = c(paste0("Total antal pat. i 2018 = ",
                                     sum(antalPatienterDat$antal)),
                                     paste0("Gnst. antal pat. pr maaned = ",
                                     round(mean(antalPatienterDat$antal),1))))

}
