% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\encoding{UTF-8}
\name{delTekstOpEfterInchLaengde}
\alias{delTekstOpEfterInchLaengde}
\title{Opdeling af tekst efter en bestemt bredde for teksten}
\usage{
delTekstOpEfterInchLaengde(teksten, maksLaengdeInch, cexPaaTeksten = 1)
}
\arguments{
\item{teksten}{en character string, En tekst, som man ønsker skal have en bestemst længde pr linje. Tekten må ikke indeholde newline'\\n'.}

\item{maksLaengdeInch}{numeric/integer, Den maksimale længde teksten må have pr linje i inches}

\item{cexPaaTeksten}{numeric/integer, Tekststørrelsen på teksten}
}
\value{
Teksten opdelt i linjer med newline ved mellemrum og med maksimal længde pr linje svarende til den angivne maksLaengdeInch.
}
\description{
En funktion til at dele en lang tekst op i flere linjer, hvor hver linje overholder en maksimallængde angivet i inches.
}
\details{
Funktionen deler teksten op efter mellemrum. Hvis der er for langt mellem mellemrum og teksten eller dele af denne ikke kan deles op, så vil funktionen melde fejl.
}
\examples{
tekstenOpdelt <- delTekstOpEfterInchLaengde(teksten = "Teksten maa ikke indeholde newline", maksLaengdeInch = 0.75)
cat(tekstenOpdelt)

}
