% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\encoding{UTF-8}
\name{enkeltLinePlot}
\alias{enkeltLinePlot}
\title{Kurve plot af en variabel fra et datasæt}
\usage{
enkeltLinePlot(
  datasaet,
  xVariablen,
  yVariablen,
  overskrift = NA,
  navnPaaXAksen = NA,
  navnPaaYAksen = NA,
  middelVariabel = NA,
  graenseVariabler = NA
)
}
\arguments{
\item{datasaet}{en data.frame, Datasættet med de variable som skal bruges til plottet.}

\item{xVariablen}{en character string, Navnet på x-variablen i datasaet}

\item{yVariablen}{en character string, Navnet på y-variablen i datasaet}

\item{overskrift}{en character vector eller NA, Overskriften på plottet. Hvis NA så vil overskriften sættes på ud fra xVariablen og yVariablen}

\item{navnPaaXAksen}{en character string eller NA, Navnet på x-aksen}

\item{navnPaaYAksen}{en character string eller NA, Navnet på y-aksen}

\item{middelVariabel}{en character string eller NA, Navnet på middelværdi-variablen. Hvis NA plottes der ingen middelværdi.}

\item{graenseVariabler}{en character vector af længden to eller NA. Navnet på grænsevariablerne. Hvis NA plottes der ingen grænseværdier.}
}
\value{
En kurve plot med eventuelt grænseværdier.
}
\description{
Kurve plot af en variabel fra et datasæt
}
\details{
Hvis x-variablen er en character-variabel i datasaet, så vil variablen laves om til en factor med levels svarende til rækkefølgen værdierne optræder i datasaet.
    Bemærk venligst at grænseværdierne skal beregnes på forhånd.
    De angivne grænseværdier vil blive defineret som øvre- og nedre grænse i plottet uafhængig af, hvilken type grænseværdier, der er angivet.
    Overskriften sættes på via. funktionen komTitelPaaPlot, der vil hermed gælde samme begrænsninger for overskrift som i denne funktion.
}
\examples{
antalPatienterDat <- data.frame("maaned" = month.name[1:12],
                                "antal" = sample(100:150, 12),
                                stringsAsFactors = FALSE)
antalPatienterDat$nedreGraense <- mean(antalPatienterDat$antal) - 3 * sd(antalPatienterDat$antal)
antalPatienterDat$oevreGraense <- mean(antalPatienterDat$antal) + 3 * sd(antalPatienterDat$antal)

enkeltLinePlot(datasaet = antalPatienterDat,
               xVariablen = "maaned",
               yVariablen = "antal",
               overskrift = "Antal patienter med x sygdom i 2018",
               navnPaaXAksen = "Maaned for besoeg paa afdelingen",
               navnPaaYAksen = "Antal patienter",
               graenseVariabler = c("nedreGraense", "oevreGraense"))

}
\seealso{
\code{\link{komTitelPaaPlot}}
}
