% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\encoding{UTF-8}
\name{flereLinePlot}
\alias{flereLinePlot}
\title{Kurve plot af flere variable fra det samme datasæt}
\usage{
flereLinePlot(
  datasaet,
  xVariablen,
  yVariablerne,
  yVariablerneNavneTilLegend = NA,
  overskrift = NA,
  navnPaaXAksen = NA,
  navnPaaYAksen = NA
)
}
\arguments{
\item{datasaet}{en data.frame, Datasættet med de variable som skal bruges til plottet}

\item{xVariablen}{en character string, Navnet på x-variablen i datasaet}

\item{yVariablerne}{en character vektor, Navnene på de y-variable der skal plottes}

\item{yVariablerneNavneTilLegend}{en character vektor af samme længde som yVariablerne, Navnene på y-variablene, som man ønsker dem til at stå i legend.
yVariablerneNavneTilLegend skal have samme rækkefølge som yVariablerne. Hvis NA så vil yVariablerne bruges i legend.}

\item{overskrift}{en character vector eller NA, Overskriften på plottet. Hvis NA så vil der ikke komme titel på plottet.}

\item{navnPaaXAksen}{en character string eller NA, Navnet på x-aksen}

\item{navnPaaYAksen}{en character string eller NA, Navnet på Y-aksen}
}
\value{
En kurve plot
}
\description{
Kurve plot af flere variable fra det samme datasæt
}
\details{
Hvis x-variablen er en character-variabel i datasaet, så vil variablen laves om til en factor med levels svarende til rækkefølgen værdierne optræder i datasaet.
    Overskriften sættes på via. funktionen komTitelPaaPlot, der vil hermed gælde samme begrænsninger for overskrift som i denne funktion.
    Funktionen bruger funktionen delTekstOpEfterInchLængde til at dele legend teksterne op for at hele teksten kan ses på grafen.
    Hvis legend er på flere end tre linjer vil den placeres udenfor på højre side af grafen.
}
\examples{
antalPatienterDat <- data.frame("maaned" = month.name[1:12],
                          "afdelingA" = sample(100:150, 12),
                          "afdelingB" = sample(50:90, 12),
                          "afdelingC" = sample(90:120, 12), stringsAsFactors = FALSE)
antalPatienterDat$samletAntal <- apply(antalPatienterDat[c("afdelingA","afdelingB","afdelingC")], 1, FUN = sum)

flereLinePlot(datasaet = antalPatienterDat,
               xVariablen = "maaned",
               yVariablerne = c("afdelingA", "afdelingB", "afdelingC"),
               yVariablerneNavneTilLegend = c("Afdeling A", "Afdeling B", "Afdeling C"),
               overskrift = "Antal patienter med x sygdom i 2018\n fordelt paa afdelinger",
               navnPaaXAksen = "Maaned for besoeg paa afdelingen",
               navnPaaYAksen = "Antal patienter")

flereLinePlot(datasaet = antalPatienterDat,
              xVariablen = "maaned",
              yVariablerne = c("afdelingA", "afdelingB", "afdelingC", "samletAntal"),
              yVariablerneNavneTilLegend = c("Afdeling A", "Afdeling B", "Afdeling C", "Antal pat. i alt"),
              overskrift = "Antal patienter med x sygdom i 2018\n fordelt paa afdelinger",
              navnPaaXAksen = "Maaned for besoeg paa afdelingen",
              navnPaaYAksen = "Antal patienter")

}
\seealso{
\code{\link{komTitelPaaPlot}} og \code{\link{delTekstOpEfterInchLaengde}}
}
