% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\encoding{UTF-8}
\name{komTitelPaaPlot}
\alias{komTitelPaaPlot}
\title{Kom titel på en plot}
\usage{
komTitelPaaPlot(titel, goerTekstenMindrePrLinje = TRUE)
}
\arguments{
\item{titel}{en character vector, Teksten som titlen skal bestå af. Se details for mere.}

\item{goerTekstenMindrePrLinje}{TRUE/FALSE, Hvis TRUE (by defualt) bliver størrelsen på titlen mindre og mindre per linje.}
}
\value{
ingen return. Titlen bliver sat på plottet i den aktive graphics device.
}
\description{
Funktion til at sætte titel på en plot.
    Funktionen sørger for at få delt titlen op i linjer efter opdelingen i string-vector og/eller newline-tegn.
    Funktionen melder fejl, hvis titlen er på mere end tre linjer.
    Funktionen tager udgangspunkt i default par-værdier.
}
\details{
Titlen må maks være på tre linjer. Teksten opdeles i linjer efter vectoren (hvert entry i titel svarer til en linje) og/eller med new-line '\\n'.
}
\examples{
plot(c(1:12), sample(100:150, 12), type = "l", xlab = "maaned nr.", ylab = "antal patienter", font.lab = 2)
komTitelPaaPlot(titel = "Antal patienter med x sygdom \n paa afdeling y\n i perioden jan. 2018 - dec. 18")

}
