% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udskrivTabelTilPdf.R
\name{plotTabel}
\alias{plotTabel}
\title{Plot tabel}
\usage{
plotTabel(tabel, tabelTitel, tema = tTema)
}
\arguments{
\item{tabel}{en data.frame, Tabellen der skal plottes}

\item{tabelTitel}{en character string, Titlen på tabellen}

\item{tema}{en liste, Tema til tabellen. Må ikke ændres med mindre man ved, hvad man gør.}
}
\value{
tabel i en plot
}
\description{
Plot tabel
}
\examples{
antalPatienterDat <- data.frame("afdeling" = c("afdelingA", "afdelingB", "afdelingC"),
                          "antalPatienter" = sample(50:150, 3))
plotTabel(tabel = antalPatienterDat, tabelTitel = "Oversigt over antal patienter med x sygdom")

}
