% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTekst.R
\encoding{UTF-8}
\name{plotTekst}
\alias{plotTekst}
\title{Plot tekst}
\usage{
plotTekst(overskrift = NA, tekst)
}
\arguments{
\item{overskrift}{en character string, Overskriften til teksten. Overskriften skal være en chracter string af længden en og på maks tre linjer opdelt med newline '\\n'.}

\item{tekst}{en character string, Teksten der skal plottes. Teksten skal være en chracter string af længen en og på makst 20 linjer opdelt med newline'\\n'.}
}
\value{
En plot med den ønskede tekst og overskrift.
}
\description{
Plot tekst
}
\examples{
plotTekst(overskrift = "Overskrift til teksten skrives her\n Eventuelt paa flere linjer\n men den maa ikke vaere paa over tre linjer.",
          tekst = paste0("Teksten kan vaere paa maks 20 linjer og med maks 100 tegn pr linje. \nOverskriften har ogsaa en begraensning paa ", 80, " tegn pr linje.",
                  "\nFunktionen melder fejl hvis ikke disse begraensninger overholdes."))

}
