% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udskrivTabelTilPdf.R
\encoding{UTF-8}
\name{udskrivTabelTilPdf}
\alias{udskrivTabelTilPdf}
\title{Udskriv tabel til pdf}
\usage{
udskrivTabelTilPdf(
  datasaet,
  filSti,
  titelTekst,
  antalRaekker = 33,
  tema = tTema,
  maksAntalTegnPrKol = 40
)
}
\arguments{
\item{datasaet}{en data.frame}

\item{filSti}{en chracter string, Stien til filen som der skal udskrives til. Strien skal ende med '.pdf' .}

\item{titelTekst}{en chracter string, Titlen på tabellen}

\item{antalRaekker}{en numeric/integer værdi, Maksimale antal rækker der skal på hver side. Default værdi er den maksimale værdi af antal rækker.
Man må ikke sætte antalRaekker over default værdien.}

\item{tema}{en liste, Tema til tabellen. Må ikke ændres med mindre man ved, hvad man gør.}

\item{maksAntalTegnPrKol}{en numeric/integer værdi, Den maksimale antal tegn pr. kolonne. Bruges i tilfælde af, at det viser sig, at tabellen er for bred til at være på siden.
Hvis tabellen er for bred, vil funktionen begrænse hvert eneste kolonne til at bestå af den makimalt antal tegn, som er angivet her.
Hvis tabellen stadig er for bred, melder funktionen fejl.}
}
\value{
En pdf.-fil med tabellen.
}
\description{
Udskriv tabel til pdf
}
\examples{
\dontrun{antalPatienterDat <- data.frame("maaned" = month.name[1:12],
                               "afdelingA" = sample(100:150, 12),
                                "afdelingB" = sample(50:90, 12),
                                "afdelingC" = sample(90:120, 12), stringsAsFactors = FALSE)
antalPatienterDat$samletAntal <- apply(antalPatienterDat[c("afdelingA","afdelingB","afdelingC")], 1, FUN = sum)

udskrivTabelTilPdf(datasaet = antalPatienterDat,
                   filSti = "antal patienter paa de tre udvalgte afdelinger.pdf",
                   titelTekst = "Antal patienter paa de udvalgte afdelinger")}
}
